rm platform specific code for Button
